Tests work... need some more work
So, the testing infrastructure now works. But there is more to do:

In particular, run_expectatoins is a switch statement that creates
relatively simple functions for running the tesst. However, it turns out
'equals' and 'contains' both require different methods to work when there
is only one value to check (against a matrix). In additio, many of these
functions share a lot in common, it may make more sense to use and S3
class with generic functions to run tests in each case, or a functor
approach like the testthat expectations.