First attempt at tol_about summary
For tol_about: Similar to ape's print.phylo, returns an invisible list. But
before doing so, prints out a summary of the tol information.

If the output of the function is directed into a variable, the information
will be saved.

Seems like this type of thing might benefit from using a class. For example,
for eneral functions like `print` and `summary` etc.

Anyway, trash it if you don't like it.