Work towards #23
This needs more work!

It's really too long to include in this existing vignette, and the text
accompanying the code needs to improve.

I think it is going to be hard to make a "snappy" summary with this
tree, it's just so big and there are so few coloured fish! I will keep
working on this branch until I come up with something worth adding to
the repo.